addded header validation

+++ app/mod_github/controllers.py
@@ -2,7 +2,7 @@ import dateutil.parser
 from ..models import Project, ProjectSchema
 from .. import db
 from . import mod_github
-from flask import jsonify
+from flask import jsonify, request, abort
 import requests
 # needed to get requests to run on GAE
 from requests_toolbelt.adapters import appengine
@@ -25,6 +25,9 @@ def get_data():
 @mod_github.route('/run')
 def run_updater():
     """ an endpoint to update the github project database, only intended to run 60 times a day using the GAE cron interface, more than this will lead to being locked out by the github API. Deletes all rows in the table and then re-populates them with the new data"""
+    # Verify request comes from GAE cron
+    if 'X-Appengine-Cron' not in request.headers:
+        abort(403)
     db.session.query(Project).delete()
     db.session.commit()
     resp = requests.get("https://api.github.com/users/bluzomby/repos")
